+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/repository/DashSpendRepository.kt
@@ -39,5 +39,5 @@ interface DashSpendRepository {
         merchantId: String
     ): GiftCardInfo
     suspend fun getGiftCard(giftCardId: String): GiftCardInfo?
-    fun getGiftCardDiscount(merchantId: String, denomination: Int): Double
+    fun getGiftCardDiscount(merchantId: String, denomination: Double): Double
 }
